+++ src/presentation/controllers/signup.spec.ts
@@ -1,6 +1,4 @@
-import { InvalidParamError } from "../errors/invalid-param-error";
-import { MissingParamError } from "../errors/missing-param-error";
-import { ServerError } from "../errors/server-error";
+import { InvalidParamError, MissingParamError, ServerError } from "../errors";
 import { EmailValidator } from "../protocols/email-validator";
 import { SignUpController } from "./signup";
